commit by raj

+++ Projects/ToDo App/lib/model/individual_chart_data.dart
@@ -1,5 +1,5 @@
 class IndividualChartData {
-  final double x; //for position
+  final int x; //for position
   final double y; //for amount
   final String day; //for day
